# src/config/openapi.yml
openapi: 3.0.3
info:
  title: |
    The Jolokia Api Server
    (Generated by `yarn build-api-doc`.)
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 0.1.2
  description: |
    This document contains a list of currently available apis that
    can be used by the **activemq-artemis-self-provisioning plugin**
    to get access to the [management api](https://github.com/apache/activemq-artemis/blob/main/docs/user-manual/management.adoc#the-management-api) of a broker instance via 
    its [jolokia](https://jolokia.org/) endpoint.

    ### How to add new apis to the src

    The api server uses [openapi](https://www.openapis.org/) to define
    its apis. All the apis are defined in the file [openapi.yml](src/config/openapi.yml).

    To add a new api first open the [openapi.yml](src/config/openapi.yml)
    and add a new api definition, for example:

    ```yaml
    /hello:
      get:
        description: hello api
        tags:
          - greeting
        operationId: hello
        parameters:
        - name: name
          required: false
          in: query
          description: The name of a caller
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DummyResponse'
      ```

    With the above section added, you defined the new api under the path `/hello`
...

...

...

...

...
    }
...
servers:
...
security:
  - bearerAuth: []
paths:
  /server/login:
    post:
      summary: Api to log in to the api server.
      description: >
        This api is used to login to the api server.
      tags:
        - security
      operationId: serverLogin
      security: [] # no security for login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                password:
                  type: string
              required:
                - userName
                - password
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerLoginResponse'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /server/logout:
    post:
      summary: Api to log out
      description: >
        This api is used to logout the current session.
      tags:
        - security
      operationId: serverLogout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyBody'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerLogoutResponse'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
  /jolokia/login:
    post:
      summary: The login api
      description: >
        This api is used to login to a jolokia endpoint.
        It tries to get the broker mbean via the joloia url
        using the parameters passed in.

          If it succeeds, it generates a [jwt token](https://jwt.io/introduction) and returns
        it back to the client. If it fails it returns a error.

          Once authenticated, the client can access the
        apis defined in this file. With each request the client
        must include a valid jwt token in a http header
        named `jolokia-session-id`.
        The src will validate the token before processing a request is
        and rejects the request if the token is not valid.
      security: [] # no security for login
      tags:
        - security
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brokerName:
                  type: string
                  description: 'identity of the broker instance, must in form of {cr-name}-{pod-ordinal}:{namespace}. For example ex-aao-0:test1'
                userName:
                  type: string
                  description: 'The user name'
                password:
                  type: string
                  description: 'The password'
                jolokiaHost:
                  type: string
                  description: 'The host name of the broker'
                scheme:
                  type: string
                  description: 'either *http* or *https*'
                port:
                  type: string
                  description: 'port number of jolokia endpoint'
              required:
                - brokerName
                - userName
                - password
                - jolokiaHost
                - scheme
                - port
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/FailureResponse'
...
components:
  schemas:
...
    ServerLogoutResponse:
      type: object
      required:
        - status
        - message
      properties:
        message:
          type: string
        status:
          type: string
    ServerLoginResponse:
      type: object
      required:
        - status
        - message
        - bearerToken
      properties:
        message:
          type: string
        status:
          type: string
        bearerToken:
          type: string
          description: The jwt token
    LoginResponse:
      type: object
      required:
        - status
        - message
        - jolokia-session-id
      properties:
        message:
          type: string
        status:
          type: string
        jolokia-session-id:
          type: string
          description: The jwt token
...
    FailureResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: [failed, error]
        message:
          type: string
...
    EmptyBody:
      type: object
      nullable: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT